requirements.txt
gspread
oauth2client
requests
beautifulsoup4
validate_email
dnspython
